feat 69 corrections

+++ app/src/main/java/edu/iesam/meceiot/core/data/local/db/converters/DateConverter.kt
@@ -18,13 +18,13 @@ class DateConverter {
     }
 
     @TypeConverter
-    fun fromFloatList(value: List<Float>?): String? {
+    fun fromFloatList(value: List<Long>?): String? {
         return Gson().toJson(value)
     }
 
     @TypeConverter
-    fun toFloatList(value: String?): List<Float>? {
-        val listType = object : TypeToken<List<Float>>() {}.type
+    fun toFloatList(value: String?): List<Long>? {
+        val listType = object : TypeToken<List<Long>>() {}.type
         return Gson().fromJson(value, listType)
     }
 }
